+++ firmware/tinyg/spindle.h
@@ -36,7 +36,7 @@
 void sp_init();
 uint8_t cm_spindle_control(uint8_t spindle_mode);
 void cm_exec_spindle_control(uint8_t spindle_mode);
-uint8_t cm_set_spindle_speed(double speed);
-void cm_exec_spindle_speed(double speed);
+uint8_t cm_set_spindle_speed(float speed);
+void cm_exec_spindle_speed(float speed);
 
 #endif
